requirements.txt
rdflib==7.0.0
PyYAML==6.0.2
pyshacl==0.24.0
pytest==8.3.3
